# platform/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_config.setRequest:
    properties:
      category:
        type: string
      description:
        type: string
      value:
        type: string
    required:
    - value
    type: object
...
  internal_dev.codeRequest:
    properties:
      model:
        type: string
      prompt:
        type: string
      repo_dir:
        type: string
    required:
    - prompt
    - repo_dir
    type: object
...
info:
  contact: {}
  description: 企业 AI 原生研发平台后端（AI 驱动 需求→研发→测试→审批→发布 全流程）
  title: 智源 ANP 平台 API
  version: "1.0"
paths:
...
  /code:
    post:
      consumes:
      - application/json
      parameters:
      - description: 编码任务(repo_dir+prompt+model)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_dev.codeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: task_id+status=running+note
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 提交失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 异步提交编码任务（立即返回 task_id）
      tags:
      - dev
  /code-tasks/{id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 查询编码任务状态/产出
      tags:
      - dev
  /config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 系统配置列表
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出全部系统配置
      tags:
      - config
  /config/{key}:
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 配置值(value)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_config.setRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{key,value}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 新增/更新一条系统配置
      tags:
      - config
...
  /project-spaces/{id}/code-tasks:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出项目空间的编码任务
      tags:
      - dev
...
  /project-spaces/{id}/usage:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用量明细列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 算力用量明细
      tags:
      - compute
  /project-spaces/{id}/usage/stats:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用量统计
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 算力用量统计看板
      tags:
      - compute
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"